4.8.1-rc1

+++ plugins/source/digitalocean/resources/plugin/client.go
@@ -79,7 +79,7 @@ func Configure(_ context.Context, logger zerolog.Logger, specBytes []byte, opts 
 		logger: logger,
 		scheduler: scheduler.NewScheduler(
 			scheduler.WithLogger(logger),
-			scheduler.WithConcurrency(uint64(config.Concurrency)),
+			scheduler.WithConcurrency(config.Concurrency),
 		),
 		tables: getTables(),
 	}, nil
